<Step5> domain / view 파일 분리
