+++ src/main/kotlin/step4/domain/Randomizer.kt
@@ -1,0 +1,10 @@
+package step4.domain
+
+object Randomizer {
+    const val RANDOM_MAX = 9
+    const val RANDOM_MIN = 0
+
+    fun getRandom(): Int {
+        return (RANDOM_MIN..RANDOM_MAX).random()
+    }
+}
