+++ src/main/kotlin/step4/main.kt
@@ -1,10 +1,15 @@
 package step4
 
+import step4.domain.RaceGame
+import step4.domain.RacingCar
+import step4.view.InputView
+import step4.view.ResultView
+
 fun main() {
     val carNames = InputView.inputCarNames()
     val tryCount = InputView.inputTryCount()
 
-    val cars = carNames.map { Car(it) }
+    val cars = carNames.map { RacingCar(it) }
 
     val game = RaceGame(cars, tryCount)
     val winners = game.playGame()
