+++ src/main/kotlin/step4/view/InputView.kt
@@ -1,4 +1,4 @@
-package step4
+package step4.view
 
 import step4.ExceptionType.ERROR_TRY_COUNT
 
@@ -16,27 +16,5 @@ object InputView {
         return readLine()
             ?.takeIf { it.matches(Regex("\\d+")) }
             ?.toInt() ?: throw IllegalArgumentException(ERROR_TRY_COUNT)
-    }
-}
-
-object ResultView {
-    private const val WINNER = "%s가 최종 우승했습니다."
-
-    fun printHistory(raceHistory: List<RoundSet>) {
-        raceHistory.forEach { roundSet ->
-            printPosition(roundSet)
-        }
-        println()
-    }
-
-    fun printPosition(roundSet: RoundSet) {
-        roundSet.carList.forEach { car ->
-            println("${car.name}: ${"-".repeat(car.position)}")
-        }
-        println()
-    }
-
-    fun printWinners(winners: String) {
-        print(WINNER.format(winners))
     }
 }
